main/forms.py
from django import forms
from django.contrib.auth import password_validation
from django.core.exceptions import ValidationError
from django.forms import inlineformset_factory
from captcha.fields import CaptchaField

from .models import user_registrated, SuperRubric, SubRubric, Anno, AdditionalImage
from .models import AnnoUser, Comment


class ChangeUserDataForm(forms.ModelForm):
    email = forms.EmailField(required=True, label='Aдpec электронной почты'),


    class Meta:
        model = AnnoUser
        fields = ('username', 'email', 'first_name', 'last_name', 'send_messages') 


class RegisterUserForm(forms.ModelForm):
    email = forms.EmailField(required=True, label='Aдpec Электронной почты')
    password1 = forms.CharField(label='Пapoль', widget=forms.PasswordInput, 
    help_text=password_validation.password_validators_help_text_html())
    password2 = forms.CharField(label='Пapoль (повторно)',  widget=forms.PasswordInput, help_text='Введите пароль еще раз')
    
    def clean_password1(self):
        password1 = self.cleaned_data['password1']

        if password1:
            password_validation.validate_password(password1)

        return password1

    def clean(self):
        super().clean()
        password1 = self.cleaned_data['password1']
        password2 = self.cleaned_data['password2']

        if password1 and password2 and password1 != password2:
            errors = {'password2': ValidationError(
            'Введенные пароли не совпадают', code='password_mismatch')}
            raise ValidationError(errors)

    def save(self, commit=True):
        user = super().save(commit=False)
        user.set_password(self.cleaned_data['password1'])
        user.is_active = False
        user.is_activated = False

        if commit:
            user.save()

        user_registrated.send(RegisterUserForm, instance=user)
        return user


    class Meta:
        model = AnnoUser
        fields = ('username', 'email', 'password1', 'password2', 'first_name', 'last_name', 'send_messages')


class SubRubricForm(forms.ModelForm):
    super_rubric = forms.ModelChoiceField(queryset=SuperRubric.objects.all(), label='Надрубрика',
    empty_label=None, required=True)
    

    class Meta:
        model = SubRubric
        fields = '__all__'


class SearchForm(forms.Form):
    keyword = forms.CharField(required=False, max_length=20, label='')


class AnnoForm(forms.ModelForm):


    class Meta:
        model = Anno
        fields = '__all__'
        widgets = { 'author': forms.HiddenInput }

AIFormSet = inlineformset_factory(Anno, AdditionalImage, fields='__all__')



class UserCommentForm(forms.ModelForm):


    class Meta:
        model = Comment
        exclude = ('is_active',)
        widgets = {'anno' : forms.HiddenInput}

class GuestCommentForm(forms.ModelForm):
    captcha = CaptchaField(label='Bвeдитe текст с картинки', error_messages={'invalid': 'Неправильный текст'})
    
    
    class Meta:
        model = Comment
        exclude = ('is_active',)
        widgets = {'anno': forms.HiddenInput} 