requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.2
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
easy-thumbnails==2.7
Pillow==7.0.0
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
